requirements.txt
langchain
langchain-community
langchain-chroma
langchain-huggingface
datasets
chromadb
python-dotenv